CLDCONNECT-2296: Devkit 3.6.0 Migration and Hadoop Client 2.6.0
1. Migrated to Anypoint Connector Devkit  3.6.0
2. Updated the Hadoop client to 2.6.0
3. Updated the demo flow and automation test cases.
4. Updated the USAGE.md and CHANGELOG.md to reflect the above changes.
