+++ .gitignore
@@ -1,18 +1,14 @@
-.DS_Store
-*~
-*.swp
-*.iml
+.studio
+target
 *.iws
 *.ipr
-*.class
-*.log
-target/
-bin/
+*.iml
+.settings
 .classpath
 .project
-.settings/
-.metadata
+.checkstyle
+.springBeans
 .idea
-Ddo*
-.mule
-
+.DS_Store
+.factorypath
+dependency-reduced-pom.xml
