+++ demo/src/main/app/mule-app.properties
@@ -1,5 +1,2 @@
-#** GENERATED CONTENT ** Mule Custom Properties file.
-#Mon Jan 14 11:20:49 GMT-03:00 2013
-
 nameNodeUri=
 sysUser=
