+++ demo/src/main/app/mule-deploy.properties
@@ -2,4 +2,4 @@
 redeployment.enabled=true
 encoding=UTF-8
 domain=default
-config.resources=Hadoop Cloud Connector Demo Project.xml
+config.resources=HDFS Connector Demo Project.xml
