+++ doc/mule-module-hdfs.xml.sample
@@ -26,90 +26,90 @@
 
 <!-- BEGIN_INCLUDE(hdfs:read-1) -->
 <!-- Reading a file using the default buffer size -->
-<hdfs:read path="/tmp/test.dat" />
+<hdfs:read path="/tmp/test.dat" config-ref="hdfs-conf"/>
 <!-- END_INCLUDE(hdfs:read-1) -->
 
 <!-- BEGIN_INCLUDE(hdfs:read-2) -->
 <!-- Reading a file using a specific buffer size -->
-<hdfs:read path="/tmp/test.dat" bufferSize="8192" />
+<hdfs:read path="/tmp/test.dat" bufferSize="8192" config-ref="hdfs-conf"/>
 <!-- END_INCLUDE(hdfs:read-2) -->
 
 <!-- BEGIN_INCLUDE(hdfs:get-metadata) -->
 <!-- Store the meta-information of a path in flow variables -->
-<hdfs:get-metadata path="/tmp/test.dat" />
+<hdfs:get-metadata path="/tmp/test.dat" config-ref="hdfs-conf"/>
 <!-- END_INCLUDE(hdfs:get-metadata) -->
 
 <!-- BEGIN_INCLUDE(hdfs:write-1) -->
 <!-- Writing to a file using the configuration -->
-<hdfs:write path="/tmp/test.dat" />
+<hdfs:write path="/tmp/test.dat" config-ref="hdfs-conf"/>
 <!-- END_INCLUDE(hdfs:write-1) -->
 
 <!-- BEGIN_INCLUDE(hdfs:write-2) -->
 <!-- Writing to a file using all the configurable parameters -->
-<hdfs:write path="/tmp/test.dat" permission="511" overwrite="true" bufferSize="8192" replication="1" blockSize="8192" ownerUserName="" ownerGroupName="" />
+<hdfs:write path="/tmp/test.dat" permission="511" overwrite="true" bufferSize="8192" replication="1" blockSize="8192" ownerUserName="" ownerGroupName="" config-ref="hdfs-conf"/>
 <!-- END_INCLUDE(hdfs:write-2) -->
 
 <!-- BEGIN_INCLUDE(hdfs:append-1) -->
 <!-- Appending to a file using the default buffer size -->
-<hdfs:append path="/tmp/test.dat" />
+<hdfs:append path="/tmp/test.dat" config-ref="hdfs-conf"/>
 <!-- END_INCLUDE(hdfs:append-1) -->
 
 <!-- BEGIN_INCLUDE(hdfs:append-2) -->
 <!-- Appending to a file using a specific buffer size -->
-<hdfs:append path="/tmp/test.dat" bufferSize="8192" />
+<hdfs:append path="/tmp/test.dat" bufferSize="8192" config-ref="hdfs-conf"/>
 <!-- END_INCLUDE(hdfs:append-2) -->
 
 <!-- BEGIN_INCLUDE(hdfs:delete-file) -->
 <!-- Delete a file -->
-<hdfs:delete-file path="/tmp/test.dat" />
+<hdfs:delete-file path="/tmp/test.dat" config-ref="hdfs-conf"/>
 <!-- END_INCLUDE(hdfs:delete-file) -->
 
 <!-- BEGIN_INCLUDE(hdfs:delete-directory) -->
 <!-- Delete a directory -->
-<hdfs:delete-directory path="/tmp/my-dir" />
+<hdfs:delete-directory path="/tmp/my-dir" config-ref="hdfs-conf"/>
 <!-- END_INCLUDE(hdfs:delete-directory) -->
 
 <!-- BEGIN_INCLUDE(hdfs:make-directories-1) -->
 <!-- Make any missing directories for the specified path, using the default file system permission -->
-<hdfs:make-directories path="/tmp/my-dir" />
+<hdfs:make-directories path="/tmp/my-dir" config-ref="hdfs-conf"/>
 <!-- END_INCLUDE(hdfs:make-directories-1) -->
 
 <!-- BEGIN_INCLUDE(hdfs:make-directories-2) -->
 <!-- Make any missing directories for the specified path, using the provided file system permission -->
-<hdfs:make-directories path="/tmp/my-dir" permission="511" />
+<hdfs:make-directories path="/tmp/my-dir" permission="511" config-ref="hdfs-conf"/>
 <!-- END_INCLUDE(hdfs:make-directories-2) -->
 
 <!-- BEGIN_INCLUDE(hdfs:rename) -->
 <!-- Rename any source directory or file to the provided target path -->
-<hdfs:rename source="/tmp/my-dir" target="/tmp/new-dir" />
+<hdfs:rename source="/tmp/my-dir" target="/tmp/new-dir" config-ref="hdfs-conf"/>
 <!-- END_INCLUDE(hdfs:rename) -->
 
 <!-- BEGIN_INCLUDE(hdfs:list-status) -->
 <!-- List the statuses of the given path -->
-<hdfs:list-status path="/tmp/my-dir" filter="^.*/2014/02/$" />
+<hdfs:list-status path="/tmp/my-dir" filter="^.*/2014/02/$" config-ref="hdfs-conf"/>
 <!-- END_INCLUDE(hdfs:list-status) -->
 
 <!-- BEGIN_INCLUDE(hdfs:glob-status) -->
 <!-- Return all the files that match file pattern, sorted by their names -->
-<hdfs:glob-status pathPattern="/tmp/*/*"/>
+<hdfs:glob-status pathPattern="/tmp/*/*" config-ref="hdfs-conf"/>
 <!-- END_INCLUDE(hdfs:glob-status) -->
 
 <!-- BEGIN_INCLUDE(hdfs:copy-from-local-file) -->
 <!-- Copy from source local disk to the target FileSystem -->
-<hdfs:copy-from-local-file deleteSource="true" overwrite="false" source="/tmp/mulesoft/" target="/user/mulesoft/" />
+<hdfs:copy-from-local-file deleteSource="true" overwrite="false" source="/tmp/mulesoft/" target="/user/mulesoft/" config-ref="hdfs-conf"/>
 <!-- END_INCLUDE(hdfs:copy-from-local-file) -->
 
 <!-- BEGIN_INCLUDE(hdfs:copy-to-local-file) -->
 <!-- Copy to source local disk from the target FileSystem -->
-<hdfs:copy-to-local-file deleteSource="false" useRawLocalFileSystem="false" source="/tmp/mulesoft/" target="/user/mulesoft/" />
+<hdfs:copy-to-local-file deleteSource="false" useRawLocalFileSystem="false" source="/tmp/mulesoft/" target="/user/mulesoft/" config-ref="hdfs-conf"/>
 <!-- END_INCLUDE(hdfs:copy-to-local-file) -->
 
 <!-- BEGIN_INCLUDE(hdfs:set-owner) -->
 <!-- Set owner of a path to change. -->
-<hdfs:set-owner path="/tmp/my-dir" ownername="mulesoft" groupname="supergroup" />
+<hdfs:set-owner path="/tmp/my-dir" ownername="mulesoft" groupname="supergroup" config-ref="hdfs-conf"/>
 <!-- END_INCLUDE(hdfs:set-owner) -->
 
 <!-- BEGIN_INCLUDE(hdfs:set-permission) -->
 <!-- Set permission of a path to change. -->
-<hdfs:set-permission path="/tmp/my-dir" permission="511" />
+<hdfs:set-permission path="/tmp/my-dir" permission="511" config-ref="hdfs-conf"/>
 <!-- END_INCLUDE(hdfs:set-permission) -->
