+++ src/test/java/org/mule/modules/hdfs/automation/testcases/CopyToLocalFileTestCases.java
@@ -1,5 +1,5 @@
 /**
- * (c) 2003-2014 MuleSoft, Inc. The software in this package is
+ * (c) 2003-2015 MuleSoft, Inc. The software in this package is
  * published under the terms of the CPAL v1.0 license, a copy of which
  * has been included with this distribution in the LICENSE.md file.
  */
@@ -15,7 +15,9 @@ import org.junit.experimental.categories.Category;
 import org.mule.modules.hdfs.automation.HDFSTestParent;
 import org.mule.modules.hdfs.automation.RegressionTests;
 import org.mule.modules.tests.ConnectorTestUtils;
+
 import java.io.File;
+
 import static org.junit.Assert.fail;
 
 @Ignore("Fails on Amazon EC2, run this test on local Hadoop instance")
@@ -48,6 +50,6 @@ public class CopyToLocalFileTestCases extends HDFSTestParent {
 
     @After
     public void tearDown() throws Exception {
-        FileUtils.cleanDirectory(new File((String) getTestRunMessageValue("target")));
+        FileUtils.deleteDirectory(new File((String) getTestRunMessageValue("target")));
     }
 }
