+++ src/test/java/org/mule/modules/hdfs/automation/testcases/GetMetadataTestCases.java
@@ -1,5 +1,5 @@
 /**
- * (c) 2003-2014 MuleSoft, Inc. The software in this package is
+ * (c) 2003-2015 MuleSoft, Inc. The software in this package is
  * published under the terms of the CPAL v1.0 license, a copy of which
  * has been included with this distribution in the LICENSE.md file.
  */
